+++ order-execution-service/src/main/java/com/kvvssut/orderexec/bean/Trade.java
@@ -1,4 +1,6 @@
 package com.kvvssut.orderexec.bean;
 
-public record Trade(String tradeId, String buyOrderId, String sellOrderId, double price, int quantity) {
+import java.math.BigDecimal;
+
+public record Trade(String tradeId, String buyOrderId, String sellOrderId, BigDecimal price, int quantity) {
 }
