+++ order-execution-service/src/main/java/com/kvvssut/orderexec/repository/OrderBook.java
@@ -4,14 +4,15 @@ import com.kvvssut.orderexec.bean.Order;
 import com.kvvssut.orderexec.bean.Side;
 import com.kvvssut.orderexec.bean.Trade;
 
+import java.math.BigDecimal;
 import java.util.*;
 import java.util.concurrent.atomic.AtomicInteger;
 
 public class OrderBook {
 
     private final AtomicInteger counter;
-    private TreeMap<Double, Deque<Order>> buyOrders;
-    private TreeMap<Double, Deque<Order>> sellOrders;
+    private TreeMap<BigDecimal, Deque<Order>> buyOrders;
+    private TreeMap<BigDecimal, Deque<Order>> sellOrders;
     private List<Trade> trades;
 
     public OrderBook() {
@@ -21,11 +22,11 @@ public class OrderBook {
         this.counter = new AtomicInteger(0);
     }
 
-    public TreeMap<Double, Deque<Order>> getOrders(Side side) {
+    public TreeMap<BigDecimal, Deque<Order>> getOrders(Side side) {
         return side == Side.BUY ? buyOrders : sellOrders;
     }
 
-    public void addTrade(String buyOrderId, String sellOrderId, double price, int quantity) {
+    public void addTrade(String buyOrderId, String sellOrderId, BigDecimal price, int quantity) {
         this.trades.add(new Trade(getAutoIncrementTradeId(), buyOrderId, sellOrderId, price, quantity));
     }
